Add Google Magritte to awesome-medipapipe
This is a new framework for face blurring developed by Google.

I didn't find a good category for this link as Magritte is not dedicated to one platform only. Please let me know if there is a better category or location in the document to add this project. Thank you!

Disclaimer: I work on the Magritte team.